% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{bunch}
\alias{bunch}
\title{Function \code{bunch}}
\usage{
bunch(from = parent.frame())
}
\arguments{
\item{from}{package (character scalar) or environment to look for operators}
}
\value{
character vector of all the available operators in an
environemnt
}
\description{
List the available operators in a package or environment.
The environment defaults to your workspace.
}
\examples{
grow(rbind, from = "base") 
bunch() # has "\%rbind\%"
myfun = function(x, y) x + y
env = new.env()
grow(myfun, to = env)
bunch() # has "\%rbind\%"
bunch(from = env) # "\%myfun\%"
}

