% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{functions}
\alias{functions}
\title{Function \code{functions}}
\usage{
functions(from = parent.frame())
}
\arguments{
\item{from}{package (character scalar) or environment to look for functions}
}
\value{
character vector of the names of available functions 
in \code{from}
}
\description{
List the available functions in a package or environment
that are not already binary ooperators.
}
\examples{
functions()
myfun = function(x, y) x + y
functions()
functions("grapes")
functions("knitr")
}
